move poke data-plane API contracts out of pages/api
