+++ front/pages/api/poke/workspaces/[wId]/data_source_views/[dsvId]/details.ts
@@ -1,19 +1,15 @@
 /** @ignoreswagger */
 // @migration-status: MIGRATED_TO_HONO
 import { withSessionAuthenticationForPoke } from "@app/lib/api/auth_wrappers";
+import type { PokeGetDataSourceViewDetails } from "@app/lib/api/poke/data_source_views";
 import { Authenticator } from "@app/lib/auth";
 import type { SessionWithUser } from "@app/lib/iam/provider";
 import { dataSourceViewToPokeJSON } from "@app/lib/poke/utils";
 import { DataSourceViewResource } from "@app/lib/resources/data_source_view_resource";
 import { apiError } from "@app/logger/withlogging";
 import type { WithAPIErrorResponse } from "@app/types/error";
-import type { PokeDataSourceViewType } from "@app/types/poke";
 import { isString } from "@app/types/shared/utils/general";
 import type { NextApiRequest, NextApiResponse } from "next";
-
-export type PokeGetDataSourceViewDetails = {
-  dataSourceView: PokeDataSourceViewType;
-};
 
 async function handler(
   req: NextApiRequest,
