+++ front/pages/api/poke/workspaces/[wId]/data_sources/[dsId]/check-stuck.ts
@@ -10,8 +10,6 @@ import { apiError } from "@app/logger/withlogging";
 import type { WithAPIErrorResponse } from "@app/types/error";
 import type { NextApiRequest, NextApiResponse } from "next";
 
-export type { CheckStuckResponseBody };
-
 async function handler(
   req: NextApiRequest,
   res: NextApiResponse<WithAPIErrorResponse<CheckStuckResponseBody>>,
