+++ front/poke/swr/data_source_details.ts
@@ -1,5 +1,5 @@
+import type { PokeGetDataSourceDetails } from "@app/lib/api/poke/data_sources";
 import { useFetcher, useSWRWithDefaults } from "@app/lib/swr/swr";
-import type { PokeGetDataSourceDetails } from "@app/pages/api/poke/workspaces/[wId]/data_sources/[dsId]/details";
 import type { LightWorkspaceType } from "@app/types/user";
 import type { Fetcher } from "swr";
