+++ front/poke/swr/data_source_view_details.ts
@@ -1,5 +1,5 @@
+import type { PokeGetDataSourceViewDetails } from "@app/lib/api/poke/data_source_views";
 import { useFetcher, useSWRWithDefaults } from "@app/lib/swr/swr";
-import type { PokeGetDataSourceViewDetails } from "@app/pages/api/poke/workspaces/[wId]/data_source_views/[dsvId]/details";
 import type { LightWorkspaceType } from "@app/types/user";
 import type { Fetcher } from "swr";
