+++ front/poke/swr/data_sources.ts
@@ -1,5 +1,5 @@
+import type { PokeListDataSources } from "@app/lib/api/poke/data_sources";
 import { emptyArray, useFetcher, useSWRWithDefaults } from "@app/lib/swr/swr";
-import type { PokeListDataSources } from "@app/pages/api/poke/workspaces/[wId]/data_sources";
 import type { PokeConditionalFetchProps } from "@app/poke/swr/types";
 import type { Fetcher } from "swr";
