+++ front/poke/swr/document.ts
@@ -1,5 +1,5 @@
+import type { PokeGetDocument } from "@app/lib/api/poke/data_sources";
 import { useFetcher, useSWRWithDefaults } from "@app/lib/swr/swr";
-import type { PokeGetDocument } from "@app/pages/api/poke/workspaces/[wId]/data_sources/[dsId]/document";
 import type { LightWorkspaceType } from "@app/types/user";
 import type { Fetcher } from "swr";
