+++ front/poke/swr/spaces.ts
@@ -1,5 +1,5 @@
+import type { PokeListSpaces } from "@app/lib/api/poke/spaces";
 import { emptyArray, useFetcher, useSWRWithDefaults } from "@app/lib/swr/swr";
-import type { PokeListSpaces } from "@app/pages/api/poke/workspaces/[wId]/spaces";
 import type { PokeConditionalFetchProps } from "@app/poke/swr/types";
 import type { Fetcher } from "swr";
